+++ src/main/java/at/hannibal2/skyhanni/features/fishing/TrophyFishMessages.kt
@@ -44,6 +44,7 @@ class TrophyFishMessages {
         if (!SkyHanniMod.feature.fishing.trophyCounter) return
 
         val message = event.message
+        //TODO replace logic with regex
         if (message.startsWith("§6§lTROPHY FISH! §r§bYou caught a")) {
             var displayName =
                 if (message.contains(" a §r")) message.between(" a §r", "§r §r") else message.between(" an §r", "§r §r")
@@ -53,7 +54,7 @@ class TrophyFishMessages {
             }
             val rarity = message.between("§r §r", "§b.").lowercase().replace("§l", "")
 
-            val name = (rarity + "_" + displayName).removeColor().lowercase().replace(" ", "")
+            val name = (rarity + "_" + displayName).removeColor().lowercase().replace(" ", "").replace("-", "")
             val amount = map.getOrDefault(name, 0) + 1
             map[name] = amount
             event.blockedReason = "trophy_fish"