+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixes
 + Fixed steaming hot flounder in custom trophy fish counter.
 + Fixed hide powder gain message from chests in crystal hollows. (thanks hypixel)
++ Fixed damage indicator for M3 Professor bug. (Thanks hypixel)
 
 ## Version 0.14